Fixed a bug that caused multiple tweets per day
LastTweetTime.txt was getting opened for writing, which wipes the file, but then never being written to. This meant that the bot would tweet every other half hour.
Also added an option to ForceTweet, which will make the bot tweet regardless of time or date. ShouldTweet still needs to be true.